pyproject.toml
[tool.poetry]
name = ""
version = "0.1.0"
description = ""
authors = ["Pavel Mikhalkevich <https://github.com/paul-michelle>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.79.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
pydantic = {extras = ["email"], version = "^1.9.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^0.20.0"
python-multipart = "^0.0.5"
python-dateutil = "^2.8.2"
SQLAlchemy = "^1.4.40"
SQLAlchemy-Utils = "^0.38.3"
psycopg2-binary = "^2.9.3"
alembic = "^1.8.1"

[tool.poetry.dev-dependencies]
pytest = "6.1.0"
pylint = "^2.14.5"
mypy = "^0.971"
requests = "^2.28.1"
types-requests = "^2.28.8"
coverage = "^6.4.4"
httpx = "^0.23.0"
pytest-trio = "^0.7.0"
isort = "^5.10.1"
sqlalchemy-stubs = "^0.4"

[tool.isort]
multi_line_output=3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"